move reducers into separate files
